move fab_location to types.py
