+++ sdk/python/packages/flet-core/src/flet_core/__init__.py
@@ -96,10 +96,7 @@ from flet_core.file_picker import (
 from flet_core.filled_button import FilledButton
 from flet_core.filled_tonal_button import FilledTonalButton
 from flet_core.flet_app import FletApp
-from flet_core.floating_action_button import (
-    FloatingActionButton,
-    FloatingActionButtonLocation,
-)
+from flet_core.floating_action_button import FloatingActionButton
 from flet_core.form_field_control import InputBorder
 from flet_core.gesture_detector import (
     DragEndEvent,
@@ -217,6 +214,7 @@ from flet_core.types import (
     BoxShape,
     ClipBehavior,
     CrossAxisAlignment,
+    FloatingActionButtonLocation,
     FontWeight,
     ImageFit,
     ImageRepeat,
